merges producer and consumer into offline persistence manager
